// src/core/types/InfoGithub.ts
export type InfoGithub = {
    id: string; 
    login:string;
    name:string;
    avatar_url: string; //--link imagem
    public_repos: string; //-- repositório publicos
    followers: string; //--seguidores
    following: string; //seguindo
    company: string; //empresa
    blog: string; // website/blog
    location: string; // Localidade
    created_at: string; // membro desde
}